Stop importing everything into toil.fileStore
