+++ src/toil/test/docs/scripts/tutorial_managing2.py
@@ -5,7 +5,7 @@ import sys
 
 def globalFileStoreJobFn(job):
     job.log("The following example exercises all the methods provided"
-            " by the toil.fileStore.FileStore class")
+            " by the toil.fileStore.fileStore.FileStore class")
 
     # Create a local temporary file.
     scratchFile = job.fileStore.getLocalTempFile()
